+++ src/apps/toolbar/modules/main/infra.tsx
@@ -3,6 +3,7 @@ import { Placeholder, ToolbarModule, ToolbarModuleType } from '../../../utils/sc
 import { DateModule } from '../Date/infra';
 import { Item } from '../item/infra';
 import { PowerModule } from '../Power/infra';
+import { SettingsModule } from '../Settings/infra';
 
 function componentByModule(module: ToolbarModule, idx: number) {
   switch (module.type) {
@@ -13,6 +14,8 @@ function componentByModule(module: ToolbarModule, idx: number) {
       return <DateModule key={idx} module={module} />;
     case ToolbarModuleType.POWER:
       return <PowerModule key={idx} module={module} />;
+    case ToolbarModuleType.SETTINGS:
+      return <SettingsModule key={idx} module={module} />;
     default:
       return null;
   }
@@ -23,8 +26,6 @@ interface Props {
 }
 
 export function ToolBar({ structure }: Props) {
-  console.log(structure);
-
   return (
     <div className="ft-bar">
       <div className="ft-bar-left">{structure.left.map(componentByModule)}</div>