+++ src/background/main.rs
@@ -17,7 +17,7 @@ mod tray;
 mod utils;
 mod windows_api;
 mod winevent;
-mod system_events;
+mod system;
 
 use cli::handle_cli_info;
 use error_handler::Result;